Adding folders to better separate Flocking and Voting, Starting this to make structure more readable going forward
