+++ SwarmCoordination/src/main/java/com/team6/swarm/intelligence/Flocking/BehaviorType.java
@@ -104,7 +104,7 @@
  * - Multiple behaviors can blend (FLOCKING + SCOUT)
  * - Priority system resolves conflicts
  */
-package com.team6.swarm.intelligence;
+package com.team6.swarm.intelligence.Flocking;
 
 public enum BehaviorType {
     /**
