+++ SwarmCoordination/src/main/java/com/team6/swarm/intelligence/Voting/VoteResult.java
@@ -119,7 +119,7 @@
  * - Logged by: Anthony's monitoring system
  * - Stored in: Recent decisions history
  */
-package com.team6.swarm.intelligence;
+package com.team6.swarm.intelligence.Voting;
 
 import java.util.Map;
 import java.util.HashMap;
